Initial implementation mapping results of partial evaluation back to query parameters for the upstream FHIR server.
The actual implementation of this is the quickest hack I could put together, but it's 'just' re-shaping one piece of data into another, so should be practical to do properly.

This system relies on de-normalizing the data from Provenance into a security label on the data point itself. While this is probably an abuse of the field, it makes filtering for 'source OR consent' very easy.